requirements.txt
Flask==3.0.3
opencv-python==4.10.0.84
numpy==1.26.4
pyyaml==6.0.2
google-generativeai==0.8.5
